% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{geom_quantileplot}
\alias{geom_quantileplot}
\title{Plot a quantile plot}
\usage{
geom_quantileplot(
  mapping = NULL,
  data = NULL,
  stat = StatQuantileplot,
  position = "dodge2",
  ...,
  outlier.colour = NULL,
  outlier.color = NULL,
  outlier.fill = NULL,
  outlier.shape = 19,
  outlier.size = 1.5,
  outlier.stroke = 0.5,
  outlier.alpha = NULL,
  quantilesP = c(0.5, 0.6, 0.7, 0.8, 0.9, 1),
  notch = FALSE,
  notchwidth = 0.5,
  varwidth = FALSE,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
Quantile plots are an alternative to box plots / histograms / violinplots as in that they visualize the distribution of data points.
I often found myself wanting a more 'robust' alternative to box plots when working with zero-inflated data (often found in -omics data).
For very sparse data, boxplots are useless as the median (and often times even upper quartile) is equal to 0/detection limit, so all
the user can see is are the outliers, which is not helpful in judging the distribution. Quantile plots are a useful alternative since they show the cutoffs of a range of quantiles.
}
